feat: implement Multi-Head Attention with GQA support (Story 3.3)
- Add Attention class with causal self-attention
- Support Group-Query Attention (GQA) with configurable K/V heads
- QK normalization applied after RoPE
- Causal masking for autoregressive generation
- Optional sliding window attention support
- 13 comprehensive tests covering MHA, GQA, and MQA configurations
- Fix RotaryEmbedding dtype consistency (Float32)
- All 38 tests pass with 0 build warnings

+++ src/NanoChat.Core/Model/RotaryEmbedding.cs
@@ -48,7 +48,7 @@ public class RotaryEmbedding : Module<Tensor, long, Tensor>
         // Shape: (dim/2,)
         var invFreq = tensor(
             Enumerable.Range(0, _dim / 2)
-                .Select(i => 1.0 / Math.Pow(_base, (2.0 * i) / _dim))
+                .Select(i => (float)(1.0 / Math.Pow(_base, (2.0 * i) / _dim)))
                 .ToArray()
         );
